+++ demo/source/demos/basicRouting/index.js
@@ -1,5 +1,4 @@
 import Root from "./Root";
-import routes from "./routes";
 import code from "./code";
 
-export { Root, routes, code };
+export { Root, code };