+++ demo/source/demos/codeSplitting/Root.js
@@ -1,13 +1,23 @@
-import React from "react";
+import React, { Component } from "react";
 
 import { StateProvider } from "jarl-react";
 
+import routes from "./routes";
 import Pages from "./Pages";
 
-const Root = ({ history, routes, basePath }) => (
-    <StateProvider history={history} routes={routes} basePath={basePath}>
-        <Pages />
-    </StateProvider>
-);
+class Root extends Component {
+    render() {
+        const { history, basePath } = this.props;
+        return (
+            <StateProvider
+                history={history}
+                routes={routes}
+                basePath={basePath}
+            >
+                <Pages />
+            </StateProvider>
+        );
+    }
+}
 
 export default Root;