+++ demo/source/demos/codeSplitting/routes.js
@@ -1,17 +1,35 @@
-// Extract dynamically loaded component into a prop called Page
-// This will end up injected into our routed props at render
+/**
+ * JARL Demos: Code Splitting
+ *
+ * This demonstrates a technique to lazy-load code splitted components
+ * during routing. By returning a Promise from a dynamic import in the
+ * `resolve` method we cause routing to wait for the component to load
+ * before completing navigation.
+ */
+
+// Map the `default` export to an object with a Page prop
 const page = ({ default: Page }) => ({ Page });
 
+// Use an artificial wait to simulate network traffic
+const wait = ms => new Promise(resolve => setTimeout(resolve, ms));
+
+// TODO: Trigger a spinner while they're loading
 const routes = [
     {
         path: "/",
         state: { page: "home" },
-        resolve: () => import("./pages/Home").then(page)
+        resolve: () =>
+            wait(500)
+                .then(() => import("./pages/Home"))
+                .then(page)
     },
     {
         path: "/big-page",
         state: { page: "bigPage" },
-        resolve: () => import("./pages/BigPage").then(page)
+        resolve: () =>
+            wait(1000)
+                .then(() => import("./pages/BigPage"))
+                .then(page)
     }
 ];