+++ demo/source/layout/MainLayout.js
@@ -1,12 +1,11 @@
 import React, { Fragment } from "react";
-import SyntaxHighlighter from "react-syntax-highlighter/prism";
-import { coy as style } from "react-syntax-highlighter/styles/prism";
 import styled from "react-emotion";
 import { Header, Label } from "semantic-ui-react";
+import { Highlight } from "./MarkdownJsx";
 
 const Grid = styled.div`
     display: grid;
-    grid-template-columns: 15rem 1fr 30rem;
+    grid-template-columns: 15rem 1.2fr 1fr;
     grid-template-rows: 8rem auto;
     padding: 3rem;
 `;
@@ -19,11 +18,19 @@ const HeaderRow = styled.div`
 const MenuSidebar = styled.div`
     padding-right: 2rem;
 `;
+
 const ContentPanel = styled.div`
     padding: 0 2rem;
 `;
+
+const FullContentPanel = styled(ContentPanel)`
+    grid-column-start: 2;
+    grid-column-end: 4;
+`;
+
 const CodePanel = styled.div`
     font-size: 80%;
+    overflow: auto;
 `;
 
 const MainLayout = ({ children, code, menu }) => (
@@ -36,18 +43,21 @@ const MainLayout = ({ children, code, menu }) => (
             <Header sub>Just Another Routing Library (for React)</Header>
         </HeaderRow>
         <MenuSidebar>{menu}</MenuSidebar>
-        <ContentPanel data-test="content">{children}</ContentPanel>
-        <CodePanel data-test="code">
-            {code &&
-                code.map(({ name, code: source }) => (
+        {code ? (
+            <ContentPanel data-test="content">{children}</ContentPanel>
+        ) : (
+            <FullContentPanel data-test="content">{children}</FullContentPanel>
+        )}
+        {code && (
+            <CodePanel data-test="code">
+                {code.map(({ name, code: source }) => (
                     <Fragment key={name}>
                         <Label>{name}:</Label>
-                        <SyntaxHighlighter language="jsx" style={style}>
-                            {source}
-                        </SyntaxHighlighter>
+                        <Highlight language="jsx" source={source} />
                     </Fragment>
                 ))}
-        </CodePanel>
+            </CodePanel>
+        )}
     </Grid>
 );