+++ demo/source/layout/MarkdownJsx.js
@@ -1,8 +1,10 @@
 import React from "react";
 import Markdown from "react-remarkable";
 import Prism from "prismjs";
+import { memoize } from "ramda";
 
 import "prismjs/components/prism-jsx";
+import "prismjs/themes/prism.css";
 import "prismjs/themes/prism-coy.css";
 
 const highlight = (code, language = "jsx") => {
@@ -20,6 +22,20 @@ const highlight = (code, language = "jsx") => {
 };
 
 const options = { highlight };
+
+const dangerousHtml = memoize((source, language = "jsx") => ({
+    __html: Prism.highlight(source, Prism.languages[language], language)
+}));
+
+export const Highlight = ({ source }) => (
+    <pre className="language-js">
+        <code
+            className="language-js"
+            // eslint-disable-next-line react/no-danger
+            dangerouslySetInnerHTML={dangerousHtml(source)}
+        />
+    </pre>
+);
 
 const MarkdownJsx = ({ source }) => (
     <Markdown options={options} source={source} />